+++ example/android/app/src/main/kotlin/live/hms/flutter/MainActivity.kt
@@ -3,10 +3,10 @@ package live.hms.flutter
 import android.app.Activity
 import android.content.Intent
 import android.content.res.Configuration
+import android.os.Build
 import android.util.Log
 import io.flutter.embedding.android.FlutterActivity
 import live.hms.hmssdk_flutter.Constants
-import live.hms.hmssdk_flutter.HmssdkFlutterPlugin
 import live.hms.hmssdk_flutter.methods.HMSPipAction
 
 class MainActivity : FlutterActivity() {
@@ -15,11 +15,13 @@ class MainActivity : FlutterActivity() {
         super.onActivityResult(requestCode, resultCode, data)
 
         if (requestCode == Constants.SCREEN_SHARE_INTENT_REQUEST_CODE && resultCode == Activity.RESULT_OK) {
-            HmssdkFlutterPlugin.hmssdkFlutterPlugin?.requestScreenShare(data)
+            data?.action = Constants.HMSSDK_RECEIVER
+            activity.sendBroadcast(data?.putExtra(Constants.METHOD_CALL, Constants.SCREEN_SHARE_REQUEST))
         }
 
         if (requestCode == Constants.AUDIO_SHARE_INTENT_REQUEST_CODE && resultCode == Activity.RESULT_OK) {
-            HmssdkFlutterPlugin.hmssdkFlutterPlugin?.requestAudioShare(data)
+            data?.action = Constants.HMSSDK_RECEIVER
+            activity.sendBroadcast(data?.putExtra(Constants.METHOD_CALL, Constants.AUDIO_SHARE_REQUEST))
         }
     }
 
@@ -34,6 +36,15 @@ class MainActivity : FlutterActivity() {
             }
         } else {
             Log.i("PIP Mode", "Exited PIP Mode")
+        }
+    }
+
+    override fun onUserLeaveHint() {
+        super.onUserLeaveHint()
+        // This should only work for android version above 8 since PIP is only supported after
+        // android 8 and will not be called after android 12 since it automatically gets handled by android.
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O && Build.VERSION.SDK_INT < Build.VERSION_CODES.S) {
+            HMSPipAction.autoEnterPipMode(this)
         }
     }
 }