+++ example/lib/common/app_dialogs/local_peer_tile_dialog.dart
@@ -1,4 +1,5 @@
 //Package imports
+
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:google_fonts/google_fonts.dart';
@@ -8,20 +9,26 @@ class LocalPeerTileDialog extends StatefulWidget {
   final String peerName;
   final bool isAudioMode;
   final bool roles;
+  final bool isVideoOn;
   final bool? isCaptureSnapshot;
   final Function() toggleCamera;
   final Function() changeName;
   final Function() changeRole;
   final Function()? captureSnapshot;
+  final Function()? localImageCapture;
+  final Function()? toggleFlash;
   const LocalPeerTileDialog(
       {required this.isAudioMode,
+      this.isVideoOn = false,
       required this.toggleCamera,
       required this.peerName,
       required this.changeRole,
       required this.roles,
       required this.changeName,
       this.captureSnapshot,
-      this.isCaptureSnapshot});
+      this.isCaptureSnapshot,
+      this.localImageCapture,
+      this.toggleFlash});
 
   @override
   _LocalPeerTileDialogState createState() => _LocalPeerTileDialogState();
@@ -41,14 +48,14 @@ class _LocalPeerTileDialogState extends State<LocalPeerTileDialog> {
         child: Column(
           mainAxisSize: MainAxisSize.min,
           children: [
-            if (!widget.isAudioMode)
-              Padding(
-                padding: const EdgeInsets.only(bottom: 20.0),
-                child: GestureDetector(
-                  onTap: () {
-                    Navigator.pop(context);
-                    widget.toggleCamera();
-                  },
+            if (!widget.isAudioMode && widget.isVideoOn)
+              GestureDetector(
+                onTap: () {
+                  Navigator.pop(context);
+                  widget.toggleCamera();
+                },
+                child: Padding(
+                  padding: const EdgeInsets.only(bottom: 20.0),
                   child: Row(
                     children: [
                       SvgPicture.asset(
@@ -66,13 +73,13 @@ class _LocalPeerTileDialogState extends State<LocalPeerTileDialog> {
                   ),
                 ),
               ),
-            Padding(
-              padding: const EdgeInsets.only(bottom: 20.0),
-              child: GestureDetector(
-                onTap: () {
-                  Navigator.pop(context);
-                  widget.changeName();
-                },
+            GestureDetector(
+              onTap: () {
+                Navigator.pop(context);
+                widget.changeName();
+              },
+              child: Padding(
+                padding: const EdgeInsets.only(bottom: 20.0),
                 child: Row(
                   children: [
                     SvgPicture.asset(
@@ -91,12 +98,12 @@ class _LocalPeerTileDialogState extends State<LocalPeerTileDialog> {
               ),
             ),
             if (widget.roles)
-              Padding(
-                padding: const EdgeInsets.only(bottom: 20.0),
-                child: GestureDetector(
-                  onTap: () {
-                    widget.changeRole();
-                  },
+              GestureDetector(
+                onTap: () {
+                  widget.changeRole();
+                },
+                child: Padding(
+                  padding: const EdgeInsets.only(bottom: 20.0),
                   child: Row(
                     children: [
                       SvgPicture.asset(
@@ -114,13 +121,37 @@ class _LocalPeerTileDialogState extends State<LocalPeerTileDialog> {
                   ),
                 ),
               ),
-            if (widget.isCaptureSnapshot != null && widget.isCaptureSnapshot!)
-              Padding(
-                padding: const EdgeInsets.only(bottom: 20.0),
-                child: GestureDetector(
-                  onTap: () {
-                    widget.captureSnapshot!();
-                  },
+            if (widget.isVideoOn)
+              GestureDetector(
+                onTap: () {
+                  Navigator.pop(context);
+                  widget.toggleFlash!();
+                },
+                child: Padding(
+                  padding: const EdgeInsets.only(bottom: 20.0),
+                  child: Row(
+                    children: [
+                      Icon(Icons.flashlight_on_outlined),
+                      SizedBox(
+                        width: 16,
+                      ),
+                      Text(
+                        "Toggle Flash",
+                        style: GoogleFonts.inter(color: iconColor),
+                      )
+                    ],
+                  ),
+                ),
+              ),
+            if (widget.isVideoOn &&
+                widget.isCaptureSnapshot != null &&
+                widget.isCaptureSnapshot!)
+              GestureDetector(
+                onTap: () {
+                  widget.captureSnapshot!();
+                },
+                child: Padding(
+                  padding: const EdgeInsets.only(bottom: 20.0),
                   child: Row(
                     children: [
                       Icon(Icons.camera_alt),
@@ -129,6 +160,27 @@ class _LocalPeerTileDialogState extends State<LocalPeerTileDialog> {
                       ),
                       Text(
                         "Capture Snapshot",
+                        style: GoogleFonts.inter(color: iconColor),
+                      )
+                    ],
+                  ),
+                ),
+              ),
+            if (widget.isVideoOn)
+              GestureDetector(
+                onTap: () {
+                  widget.localImageCapture!();
+                },
+                child: Padding(
+                  padding: const EdgeInsets.only(bottom: 20.0),
+                  child: Row(
+                    children: [
+                      Icon(Icons.image_outlined),
+                      SizedBox(
+                        width: 16,
+                      ),
+                      Text(
+                        "Local Image Capture",
                         style: GoogleFonts.inter(color: iconColor),
                       )
                     ],