+++ example/lib/meeting/meeting_store.dart
@@ -46,8 +46,6 @@ class MeetingStore extends ChangeNotifier
     _hmsSDKInteractor = hmsSDKInteractor;
   }
 
-  // HMSLogListener
-
   bool isSpeakerOn = true;
 
   int screenShareCount = 0;
@@ -174,7 +172,7 @@ class MeetingStore extends ChangeNotifier
 
   bool isPipActive = false;
 
-  bool isPipAutoEnabled = false;
+  bool isPipAutoEnabled = true;
 
   bool lastVideoStatus = false;
 
@@ -184,7 +182,11 @@ class MeetingStore extends ChangeNotifier
 
   HMSVideoTrack? currentPIPtrack;
 
-  HMSLogList? applicationLogs;
+  HMSLogList applicationLogs = HMSLogList(hmsLog: []);
+
+  bool isHLSPlayerRequired = true;
+
+  bool isFlashOn = false;
 
   Future<HMSException?> join(String userName, String roomUrl,
       {HMSConfig? roomConfig}) async {
@@ -247,7 +249,7 @@ class MeetingStore extends ChangeNotifier
     WidgetsBinding.instance.removeObserver(this);
     hmsException = null;
     if ((localPeer?.role.name.contains("hls-") ?? false) && hasHlsStarted) {
-      hlsVideoController!.dispose(forceDispose: true);
+      hlsVideoController?.dispose(forceDispose: true);
       hlsVideoController = null;
     }
     _hmsSDKInteractor.leave(hmsActionResultListener: this);
@@ -560,11 +562,15 @@ class MeetingStore extends ChangeNotifier
     getAudioDevicesList();
     notifyListeners();
 
-    if (Platform.isIOS && !(isHLSLink)) {
-      HMSIOSPIPController.setup(
-          autoEnterPip: true,
-          aspectRatio: [9, 16],
-          backgroundColor: Colors.black);
+    if (!(isHLSLink)) {
+      if (Platform.isIOS) {
+        HMSIOSPIPController.setup(
+            autoEnterPip: true,
+            aspectRatio: [9, 16],
+            backgroundColor: Colors.black);
+      } else if (Platform.isAndroid) {
+        HMSAndroidPIPController.setup();
+      }
     }
 
     FlutterForegroundTask.startService(
@@ -692,7 +698,7 @@ class MeetingStore extends ChangeNotifier
 
   @override
   void onHMSError({required HMSException error}) {
-    log("onHMSError-> error: ${error.message}");
+    log("onHMSError-> error: ${error.code} ${error.message}");
     this.hmsException = error;
     Utilities.showNotification(error.message ?? "", "error");
     notifyListeners();
@@ -986,7 +992,13 @@ class MeetingStore extends ChangeNotifier
     HMSLogList? _logsDump = await _hmsSDKInteractor.getAllogs();
     await deleteFile();
     writeLogs(_logsDump);
-    applicationLogs = null;
+    hlsVideoController?.dispose(forceDispose: true);
+    hlsVideoController = null;
+    if (Platform.isAndroid) {
+      HMSAndroidPIPController.destroy();
+    } else if (Platform.isIOS) {
+      HMSIOSPIPController.destroy();
+    }
     _hmsSDKInteractor.removeHMSLogger();
     _hmsSDKInteractor.destroy();
     peerTracks.clear();
@@ -1103,7 +1115,7 @@ class MeetingStore extends ChangeNotifier
         if (peer.isLocal) {
           localPeer = peer;
           if (hlsVideoController != null && !peer.role.name.contains("hls-")) {
-            hlsVideoController!.dispose(forceDispose: true);
+            hlsVideoController?.dispose(forceDispose: true);
             hlsVideoController = null;
           }
         }
@@ -1117,9 +1129,11 @@ class MeetingStore extends ChangeNotifier
           }
         }
 
-        // Setup or destroy PIP controller on role Based
-        if (Platform.isIOS) {
-          if (peer.isLocal) {
+        // Setup or destroy PIP controller on role basis
+        //Pip is not supported for player by default
+        //So we need to destroy the pip config
+        if (peer.isLocal) {
+          if (Platform.isIOS) {
             if (peer.role.name.contains("hls-")) {
               HMSIOSPIPController.destroy();
             } else {
@@ -1127,6 +1141,13 @@ class MeetingStore extends ChangeNotifier
                   autoEnterPip: true,
                   aspectRatio: [9, 16],
                   backgroundColor: Colors.black);
+            }
+          } else {
+            if (peer.role.name.contains("hls-")) {
+              HMSAndroidPIPController.destroy();
+            } else {
+              HMSAndroidPIPController.setup(
+                  autoEnterPip: true, aspectRatio: [9, 16]);
             }
           }
         }
@@ -1425,8 +1446,7 @@ class MeetingStore extends ChangeNotifier
       bool _isPipAvailable = await HMSAndroidPIPController.isAvailable();
       if (_isPipAvailable) {
         //[isPipActive] method can also be used to check whether application is in pip Mode or not
-        isPipActive =
-            await HMSAndroidPIPController.start(autoEnterPip: isPipAutoEnabled);
+        isPipActive = await HMSAndroidPIPController.start();
         notifyListeners();
       }
     }
@@ -1482,16 +1502,10 @@ class MeetingStore extends ChangeNotifier
     }
   }
 
-  void destroyPIPSetupOnIOS() {
-    if (Platform.isIOS) {
-      HMSIOSPIPController.destroy();
-    }
-  }
-
   void setPIPVideoController(bool reinitialise,
       {double? aspectRatio, String? hlsStreamUrl}) {
     if (hlsVideoController != null) {
-      hlsVideoController!.dispose(forceDispose: true);
+      hlsVideoController?.dispose(forceDispose: true);
       hlsVideoController = null;
     }
     if (aspectRatio != null) {
@@ -1540,9 +1554,9 @@ class MeetingStore extends ChangeNotifier
         liveStream: true);
     hlsVideoController =
         PipFlutterPlayerController(pipFlutterPlayerConfiguration);
-    hlsVideoController!.setupDataSource(dataSource);
-    hlsVideoController!.play();
-    hlsVideoController!.setPipFlutterPlayerGlobalKey(pipFlutterPlayerKey);
+    hlsVideoController?.setupDataSource(dataSource);
+    hlsVideoController?.play();
+    hlsVideoController?.setPipFlutterPlayerGlobalKey(pipFlutterPlayerKey);
     if (reinitialise) notifyListeners();
   }
 
@@ -1560,6 +1574,21 @@ class MeetingStore extends ChangeNotifier
       peerTracks.add(peerTrackNode);
     }
     notifyListeners();
+  }
+
+  ///
+  /// This method is used to toggle the flash light of your phone
+  /// Here we are not checking whether flash is supported or not
+  /// Since this method already check that internally,
+  ///
+  void toggleFlash() async {
+    dynamic result = await HMSCameraControls.toggleFlash();
+    if (result is HMSException) {
+      Utilities.showToast(
+          "Error Occured: code: ${result.code?.errorCode}, description: ${result.description}, message: ${result.message}",
+          time: 5);
+      return;
+    }
   }
 
 //Get onSuccess or onException callbacks for HMSActionResultListenerMethod
@@ -1791,6 +1820,7 @@ class MeetingStore extends ChangeNotifier
       return;
     }
     if (state == AppLifecycleState.resumed) {
+      isHLSPlayerRequired = true;
       if (Platform.isAndroid) {
         isPipActive = await HMSAndroidPIPController.isActive();
       } else if (Platform.isIOS) {
@@ -1824,6 +1854,12 @@ class MeetingStore extends ChangeNotifier
         lastVideoStatus = true;
       }
 
+      if (Platform.isAndroid) {
+        isPipActive = await HMSAndroidPIPController.isActive();
+        isHLSPlayerRequired = false;
+        notifyListeners();
+      }
+
       if (Platform.isIOS) {
         if (screenShareCount == 0 || isScreenShareOn) {
           int peerIndex = peerTracks.indexWhere((element) =>
@@ -1848,9 +1884,15 @@ class MeetingStore extends ChangeNotifier
         }
       }
     } else if (state == AppLifecycleState.inactive) {
-      if (Platform.isAndroid && !isPipActive && !isHLSLink) {
-        HMSAndroidPIPController.start(autoEnterPip: isPipAutoEnabled);
-        isPipActive = true;
+      if (Platform.isAndroid && !isPipActive) {
+        isPipActive = await HMSAndroidPIPController.isActive();
+        isHLSPlayerRequired = false;
+      }
+      notifyListeners();
+    } else if (state == AppLifecycleState.detached) {
+      if (Platform.isAndroid && !isPipActive) {
+        isPipActive = await HMSAndroidPIPController.isActive();
+        isHLSPlayerRequired = false;
       }
       notifyListeners();
     }
@@ -1861,7 +1903,7 @@ class MeetingStore extends ChangeNotifier
     FirebaseCrashlytics.instance.log(hmsLogList.toString());
     FirebaseAnalytics.instance.logEvent(
         name: "SDK_Logs", parameters: {"data": hmsLogList.toString()});
-    applicationLogs = hmsLogList;
+    applicationLogs.hmsLog.addAll(hmsLogList.hmsLog);
     notifyListeners();
   }
 }