+++ release-apps.sh
@@ -4,44 +4,68 @@
 set -e
 set -x
 
-echo "🌳🍀 git branch: $(git rev-parse --abbrev-ref HEAD)"
+perform_pub_actions() {
+	echo "🌳🍀 git branch: $(git rev-parse --abbrev-ref HEAD)"
 
-git pull --verbose
+	git pull --verbose
 
-flutter pub get
+	flutter pub get
 
-cd ./example
+	cd ./example
+}
 
-cd ./android
+release_android() {
+	cd ./android
 
-bundle install --verbose
+	bundle install --verbose
 
-bundle exec fastlane release_on_firebase
+	bundle exec fastlane release_on_firebase
+}
 
-cd ../ios
+release_iOS() {
+	cd ./ios
 
-pod install --verbose 
+	pod install --verbose
 
-bundle install --verbose
+	bundle install --verbose
 
-bundle exec fastlane distribute_app
+	bundle exec fastlane distribute_app
+}
 
-cd .. ; cd ..
+perform_git_actions() {
+	cd ..
 
-while read line; do
-    if [[ $line =~ ^versionCode.[0-9]+$ ]]; then 
-        buildNumber=$(echo $line | grep -o -E '[0-9]+')
-    elif [[ $line =~ ^versionName.*$ ]]; then
-        versionCode=$(echo $line | grep -o -E '[0-9].[0-9].[0-9]+')
-    fi
-done <example/android/app/build.gradle
+	while read line; do
+		if [[ $line =~ ^versionCode.[0-9]+$ ]]; then
+			buildNumber=$(echo $line | grep -o -E '[0-9]+')
+		elif [[ $line =~ ^versionName.*$ ]]; then
+			versionCode=$(echo $line | grep -o -E '[0-9].[0-9].[0-9]+')
+		fi
+	done <example/android/app/build.gradle
 
+	git add example/android/app/build.gradle
+	git add example/ios/Podfile.lock
+	git add example/ios/Runner/Info.plist
+	git add example/ios/Runner.xcodeproj/project.pbxproj
 
-git add example/android/app/build.gradle
-git add example/ios/Podfile.lock
-git add example/ios/Runner/Info.plist
-git add example/ios/Runner.xcodeproj/project.pbxproj
+	git commit -m "released sample app version $versionCode ($buildNumber) 🍀"
 
-git commit -m "released sample app version $versionCode ($buildNumber) 🍀"
+	git push --verbose
+}
 
-git push --verbose
+perform_pub_actions
+P1=$!
+
+wait $P1
+
+release_android &
+P2=$!
+
+release_iOS &
+P3=$!
+
+wait $P2 $P3
+
+perform_git_actions
+
+say done